+++ hail/src/main/scala/is/hail/expr/types/physical/PBaseStruct.scala
@@ -138,222 +138,40 @@ abstract class PBaseStruct extends PType {
 
   def nMissing: Int
 
-  def nMissingBytes: Int
-
   def missingIdx: Array[Int]
 
-  def byteOffsets: Array[Long]
+  def allocate(region: Region): Long
 
-  def allocate(region: Region): Long = {
-    region.allocate(alignment, byteSize)
-  }
+  def allocate(region: Code[Region]): Code[Long]
 
-  def allocate(region: Code[Region]): Code[Long] =
-    region.allocate(alignment, byteSize)
+  def initialize(structAddress: Long, setMissing: Boolean = false): Unit
 
-  def initialize(structAddress: Long, setMissing: Boolean = false): Unit = {
-    if (allFieldsRequired) {
-      return
-    }
+  def stagedInitialize(structAddress: Code[Long], setMissing: Boolean = false): Code[Unit]
 
-    Region.setMemory(structAddress, nMissingBytes.toLong, if (setMissing) 0xFF.toByte else 0.toByte)
-  }
-
-  def stagedInitialize(structAddress: Code[Long], setMissing: Boolean = false): Code[Unit] = {
-    if (allFieldsRequired) {
-      return Code._empty
-    }
-
-    Region.setMemory(structAddress, const(nMissingBytes.toLong), const(if (setMissing) 0xFF.toByte else 0.toByte))
-  }
-
-  def isFieldDefined(offset: Long, fieldIdx: Int): Boolean =
-    fieldRequired(fieldIdx) || !Region.loadBit(offset, missingIdx(fieldIdx))
+  def isFieldDefined(offset: Long, fieldIdx: Int): Boolean
 
   def isFieldMissing(off: Long, fieldIdx: Int): Boolean = !isFieldDefined(off, fieldIdx)
 
-  def isFieldMissing(offset: Code[Long], fieldIdx: Int): Code[Boolean] =
-    if (fieldRequired(fieldIdx))
-      false
-    else
-      Region.loadBit(offset, missingIdx(fieldIdx).toLong)
+  def isFieldMissing(offset: Code[Long], fieldIdx: Int): Code[Boolean]
 
   def isFieldDefined(offset: Code[Long], fieldIdx: Int): Code[Boolean] =
     !isFieldMissing(offset, fieldIdx)
 
-  def setFieldMissing(offset: Long, fieldIdx: Int) {
-    assert(!fieldRequired(fieldIdx))
-    Region.setBit(offset, missingIdx(fieldIdx))
-  }
+  def setFieldMissing(offset: Long, fieldIdx: Int): Unit
 
-  def setFieldMissing(offset: Code[Long], fieldIdx: Int): Code[Unit] = {
-    assert(!fieldRequired(fieldIdx))
-    Region.setBit(offset, missingIdx(fieldIdx).toLong)
-  }
+  def setFieldMissing(offset: Code[Long], fieldIdx: Int): Code[Unit]
 
-  def setFieldPresent(offset: Long, fieldIdx: Int) {
-    assert(!fieldRequired(fieldIdx))
-    Region.clearBit(offset, missingIdx(fieldIdx))
-  }
+  def setFieldPresent(offset: Long, fieldIdx: Int): Unit
 
-  def setFieldPresent(offset: Code[Long], fieldIdx: Int): Code[Unit] = {
-    assert(!fieldRequired(fieldIdx))
-    Region.clearBit(offset, missingIdx(fieldIdx).toLong)
-  }
+  def setFieldPresent(offset: Code[Long], fieldIdx: Int): Code[Unit]
 
-  def fieldOffset(structAddress: Long, fieldIdx: Int): Long =
-    structAddress + byteOffsets(fieldIdx)
+  def fieldOffset(structAddress: Long, fieldIdx: Int): Long
 
-  def fieldOffset(structAddress: Code[Long], fieldIdx: Int): Code[Long] =
-    structAddress + byteOffsets(fieldIdx)
+  def fieldOffset(structAddress: Code[Long], fieldIdx: Int): Code[Long]
 
-  def loadField(offset: Long, fieldIdx: Int): Long = {
-    val off = fieldOffset(offset, fieldIdx)
-    types(fieldIdx).fundamentalType match {
-      case _: PArray | _: PBinary => Region.loadAddress(off)
-      case _ => off
-    }
-  }
+  def loadField(offset: Long, fieldIdx: Int): Long
 
-  def loadField(offset: Code[Long], fieldIdx: Int): Code[Long] =
-    loadField(fieldOffset(offset, fieldIdx), types(fieldIdx))
-
-  private def loadField(fieldOffset: Code[Long], fieldType: PType): Code[Long] = {
-    fieldType.fundamentalType match {
-      case _: PArray | _: PBinary => Region.loadAddress(fieldOffset)
-      case _ => fieldOffset
-    }
-  }
-
-  def deepPointerCopy(mb: MethodBuilder, region: Code[Region], dstStructAddress: Code[Long]): Code[Unit] = {
-    var c: Code[Unit] = Code._empty
-
-    var i = 0
-    while(i < this.size) {
-      val dstFieldType = this.fields(i).typ.fundamentalType
-      if(dstFieldType.containsPointers) {
-        val dstFieldAddress = mb.newField[Long]
-        c = Code(
-          c,
-          this.isFieldDefined(dstStructAddress, i).orEmpty(
-            Code(
-              dstFieldAddress := this.fieldOffset(dstStructAddress, i),
-              dstFieldType match {
-                case t@(_: PBinary | _: PArray) =>
-                  Region.storeAddress(dstFieldAddress, t.copyFromType(mb, region, dstFieldType, Region.loadAddress(dstFieldAddress), forceDeep = true))
-                case t: PBaseStruct =>
-                  t.deepPointerCopy(mb, region, dstFieldAddress)
-                case t: PType =>
-                  fatal(s"Field type isn't supported ${t}")
-              }
-            )
-          )
-        )
-      }
-      i += 1
-    }
-
-    c
-  }
-
-  def deepPointerCopy(region: Region, dstStructAddress: Long) {
-    var i = 0
-    while(i < this.size) {
-      val dstFieldType = this.fields(i).typ.fundamentalType
-      if(dstFieldType.containsPointers && this.isFieldDefined(dstStructAddress, i)) {
-        val dstFieldAddress = this.fieldOffset(dstStructAddress, i)
-        dstFieldType match {
-          case t@(_: PBinary | _: PArray) =>
-            Region.storeAddress(dstFieldAddress, t.copyFromType(region, dstFieldType, Region.loadAddress(dstFieldAddress), forceDeep = true))
-          case t: PBaseStruct =>
-            t.deepPointerCopy(region, dstFieldAddress)
-          case t: PType =>
-            fatal(s"Field type isn't supported ${t}")
-        }
-      }
-      i += 1
-    }
-  }
-
-  def copyFromType(mb: MethodBuilder, region: Code[Region], srcPType: PType, srcStructAddress: Code[Long], forceDeep: Boolean): Code[Long] = {
-    val sourceType = srcPType.asInstanceOf[PBaseStruct]
-    assert(sourceType.size == this.size)
-
-    if (this == sourceType && !forceDeep)
-      srcStructAddress
-    else {
-      val addr = mb.newLocal[Long]
-          Code(
-            addr := allocate(region),
-            constructAtAddress(mb, addr, region, sourceType, srcStructAddress, forceDeep),
-            addr
-          )
-    }
-  }
-
-  def copyFromTypeAndStackValue(mb: MethodBuilder, region: Code[Region], srcPType: PType, stackValue: Code[_], forceDeep: Boolean): Code[_] =
-    this.copyFromType(mb, region, srcPType, stackValue.asInstanceOf[Code[Long]], forceDeep)
-
-  def copyFromType(region: Region, srcPType: PType, srcStructAddress: Long, forceDeep: Boolean): Long = {
-    val sourceType = srcPType.asInstanceOf[PBaseStruct]
-    if (this == sourceType && !forceDeep)
-      srcStructAddress
-    else {
-      val newAddr = allocate(region)
-      constructAtAddress(newAddr, region, sourceType, srcStructAddress, forceDeep)
-      newAddr
-    }
-  }
-
-  def constructAtAddress(mb: MethodBuilder, addr: Code[Long], region: Code[Region], srcPType: PType, srcAddress: Code[Long], forceDeep: Boolean): Code[Unit] = {
-    val srcStruct = srcPType.asInstanceOf[PBaseStruct]
-    val addrVar = mb.newLocal[Long]
-    if (srcStruct == this) {
-      var c: Code[Unit] = Code(
-        addrVar := addr,
-        Region.copyFrom(srcAddress, addrVar, byteSize))
-      if (forceDeep) {
-        c = Code(c, deepPointerCopy(mb, region, addrVar))
-      }
-      c
-    } else {
-      val srcAddrVar = mb.newLocal[Long]
-
-      Code(
-        srcAddrVar := srcAddress,
-        addrVar := addr,
-        stagedInitialize(addrVar, setMissing = true),
-        Code(fields.zip(srcStruct.fields).map { case (dest, src) =>
-          assert(dest.typ.required <= src.typ.required)
-          val idx = dest.index
-          assert(idx == src.index)
-          srcStruct.isFieldDefined(srcAddrVar, idx).orEmpty(
-            dest.typ.constructAtAddress(mb, fieldOffset(addrVar, idx), region, src.typ, srcStruct.loadField(srcAddrVar, idx), forceDeep))
-        }: _*
-      ).asInstanceOf[Code[Unit]])
-    }
-  }
-
-  def constructAtAddress(addr: Long, region: Region, srcPType: PType, srcAddress: Long, forceDeep: Boolean): Unit = {
-    val srcStruct = srcPType.asInstanceOf[PBaseStruct]
-    if (srcStruct == this) {
-      Region.copyFrom(srcAddress, addr, byteSize)
-      if (forceDeep)
-        deepPointerCopy(region, addr)
-    } else {
-      initialize(addr, setMissing = true)
-      var idx = 0
-      while (idx < types.length) {
-        idx += 1
-        val dest = types(idx)
-        val src = srcStruct.types(idx)
-        assert(dest.required <= src.required)
-
-        if (srcStruct.isFieldDefined(srcAddress, idx))
-          dest.constructAtAddress(fieldOffset(addr, idx), region, src, srcStruct.loadField(srcAddress, idx), forceDeep)
-      }
-    }
-  }
+  def loadField(offset: Code[Long], fieldIdx: Int): Code[Long]
 
   override def containsPointers: Boolean = types.exists(_.containsPointers)
 }
