+++ hail/src/main/scala/is/hail/expr/types/physical/PStruct.scala
@@ -28,7 +28,7 @@ object PStruct {
   def canonical(t: PType): PStruct = PCanonicalStruct.canonical(t)
 }
 
-abstract class PStruct extends PBaseStruct {
+trait PStruct extends PBaseStruct {
   lazy val virtualType: TStruct = TStruct(fields.map(f => Field(f.name, f.typ.virtualType, f.index)), required)
 
   def copy(fields: IndexedSeq[PField] = this.fields,  required: Boolean = this.required): PStruct
